requirements.txt
fastapi
uvicorn[standard]
transformers
torch
Pillow
python-multipart
numpy<2
